+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/domain/base/service/AbstractAssetBaseService.java
@@ -26,6 +26,7 @@ import org.eclipse.tractusx.traceability.assets.domain.base.model.AssetBase;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.Owner;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.QualityType;
 import org.eclipse.tractusx.traceability.assets.infrastructure.asbuilt.model.ManufacturingInfo;
+import org.eclipse.tractusx.traceability.common.model.SearchCriteria;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.BomLifecycle;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.Direction;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.relationship.Aspect;
@@ -120,6 +121,10 @@ public abstract class AbstractAssetBaseService implements AssetBaseService {
 
     public PageResult<AssetBase> getAssets(Pageable pageable, Owner owner) {
         return getAssetRepository().getAssets(pageable, owner);
+    }
+
+    public PageResult<AssetBase> getAssets(Pageable pageable, List<SearchCriteria> filter) {
+        return getAssetRepository().getAssets(pageable, filter);
     }
 
     public AssetBase getAssetById(String assetId) {